Added code for creating backend servers
Added logic for creating the backend servers.
Small change in logic for creating the hosts and
config file. They will now only add the names that are
isntances created with.
Also when server exists before the creation of server
used to return none when there was a server exists.
But now the return will send the same while creating
the server. So there all are placed in the instances.json file
and used in creating hosts and config files.