Simplified the install file
Simplified the installl file so that creation of server and
running of playbook will be easy. Also added a new function,
it checks if all server are reachable using ansible ping.
For every 5 secs it checks and 6 attempts are made total.
If servers are not reachable after six attempts it exits the
script.
The name of cinstall.py is changed to install.